requirements.txt
python-dotenv
flask
unidecode
waitress